Bug fix
Fixed a bug where onMouseUp event would only stop the selection of items (in ItemDisplayComponent) if the event would fire over one of the items. This was because the onMouseUp event was registered on the item itself. It is now registered on ScalesContainer, and it dispatches a generic MOUSE_UP action. It is possible to refactor it later, so that it is registered on the root element always.